refactor: move to vite bundler
