+++ package.json
@@ -2,21 +2,24 @@
   "name": "react-todo-app",
   "version": "0.1.0",
   "private": true,
-  "devDependencies": {
-    "react-scripts": "0.9.0"
-  },
+  "type": "module",
   "dependencies": {
     "bootstrap": "^3.4.1",
     "immutability-helper": "^2.1.1",
-    "keycode-js": "^0.0.4",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "react": "^15.7.0",
+    "react-dom": "^15.7.0",
     "recompose": "^0.23.5"
   },
   "scripts": {
     "start": "react-scripts start",
+    "dev": "vite",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.2.1",
+    "react-scripts": "^0.9.0",
+    "vite": "^5.1.6"
   }
 }
