+++ src/util/common.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 /**
  * Returns a new object with only few attributes of the original object.
@@ -9,20 +9,22 @@ import React from 'react';
  * @return {Object}            New object with only the selected attributes.
  */
 export function objectWithOnly(object, attrs) {
-    let newObject = {};
+  let newObject = {};
 
-    attrs.forEach(attr => {
-        newObject[attr] = object[attr].bind(object);
-    });
+  attrs.forEach((attr) => {
+    newObject[attr] = object[attr].bind(object);
+  });
 
-    return newObject;
+  return newObject;
 }
 
 /**
  * Wraps react children elements with props.
  */
 export function wrapChildrenWith(children, props) {
-    return React.Children.map(children, child => React.cloneElement(child, props));
+  return React.Children.map(children, (child) =>
+    React.cloneElement(child, props)
+  );
 }
 
 /**
@@ -33,5 +35,5 @@ export function wrapChildrenWith(children, props) {
  * @return {Boolean}
  */
 export function stringInclues(str, substr) {
-    return str.indexOf(substr) !== -1;
+  return str.indexOf(substr) !== -1;
 }
